chore(web): split agents workbench panel state
